CDC-15 Refactor: lib/wishlist.ts
